test(core): split runtime shadowing tests
